+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Upcoming...
 * ... <!-- Add new lines here. -->
 * fix: Errors when handling sibling `@at-root` rule blocks
+* fix: Move all preceeding comments with rule
 
 
 ## 5.0.6